ShortestPath/src/main/java/br/com/bexs/shortestpath/Route.java
package br.com.bexs.shortestpath;

public class Route {

	private String origin;
	private String destination;
	private int distance;
	
	public Route() { }
	
	public Route(String origin, String destination, int distance) {
		this.origin = origin;
		this.destination = destination;
		this.distance = distance;
	}
	
	public String getOrigin() {
		return origin;
	}
	public void setOrigin(String origin) {
		this.origin = origin;
	}
	public String getDestination() {
		return destination;
	}
	public void setDestination(String destination) {
		this.destination = destination;
	}
	public int getDistance() {
		return distance;
	}
	public void setDistance(int distance) {
		this.distance = distance;
	}
}
